move state into its own file
